cmake: Fix libbtbb.pc.in.
The @prefix@, @exec_prefix@, @libdir@ and @includedir@ variables are not
set by cmake. Set the values in the libbtbb.pc.in file by including the
GNUInstallDirs.cmake file in CMakeLists.txt. This creates the
CMAKE_INSTALL_LIBDIR and CMAKE_INSTALL_INCLUDEDIR variables.

This change allows user to configure for installing into non-standard
locations as such:

    mkdir build
    cd build
    cmake -DCMAKE_INSTALL_PREFIX=$HOME/local ..