simplify btbb_decode()
It makes no sense to iterate over CLK6 here. If CLK6 is not correct here, we are in trouble anyway.

The part which was commented checks if the channel of the packet matches the hopping pattern. This brings no benefit since when following the Ubertooth stick is only able to receive packets that match the hopping pattern.